backend/pyproject.toml
[project]
name = "medical-messenger-platform"
version = "0.1.0"
description = "통합 미용의료 AI 상담 플랫폼"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "gunicorn>=22.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",
    # Redis
    "redis>=5.0.0",
    # Auth
    "pyjwt>=2.9.0",
    "pwdlib[argon2]>=0.2.0",
    # Validation & Settings
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "email-validator>=2.2.0",
    # Celery
    "celery[rabbitmq]>=5.4.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Monitoring
    "prometheus-client>=0.21.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    # HTTP client
    "httpx>=0.27.0",
    # File handling
    "python-multipart>=0.0.12",
    "openpyxl>=3.1.0",
    # LangChain (Phase 2)
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "langsmith>=0.2.0",
    # Azure
    "azure-storage-blob>=12.0.0",
    # Resilience
    "tenacity>=9.0.0",
    # Payments
    "stripe>=8.0.0",
    # PDF generation
    "reportlab>=4.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
